refactor: rename editor-provider to editor-context
Renamed `editor-provider` and related imports, exports, and references to `editor-context` for consistency with naming conventions. Updated all affected editor components to reflect the changes.
