+++ apps/nextjs/src/components/ui/editor/editors/editor-placeholder.tsx
@@ -2,7 +2,7 @@ import { memo, useMemo } from 'react';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { Text } from '@/components/ui/text';
 import { isContentEmpty } from '@/shared/prosemirror/utils';
-import { useEditorViewContext } from './editor-provider';
+import { useEditorViewContext } from './editor-context';
 
 type Props = FlexProps;
 export const EditorPlaceholder = memo(function EditorPlaceholder(props: Props) {
