+++ apps/nextjs/src/components/ui/editor/editors/index.ts
@@ -1,5 +1,5 @@
 export * from './editor-container';
 export * from './editor-content';
-export * from './editor-provider';
+export * from './editor-context';
 export * from './react-node-view';
 export * from './react-node-view-context';
